+++ lib/n_request.dart
@@ -1,11 +1,10 @@
 library n_request;
 
-import 'dart:typed_data';
-
-import 'package:http/http.dart';
-import 'package:n_request/src/enums.dart';
-import 'package:n_request/src/models.dart';
-import 'package:n_request/src/request.dart';
+import 'dart:typed_data' show Uint8List;
+import 'package:http/http.dart' show MultipartFile;
+import 'package:n_request/src/enums.dart' show RequestType;
+import 'package:n_request/src/models.dart' show ResponseData;
+import 'package:n_request/src/request.dart' show NCustomRequest;
 
 export '/src/enums.dart';
 export '/src/models.dart';
@@ -19,19 +18,27 @@ class NRequest{
   final Map<String, dynamic>  body;
   final List<MultipartFile>   files;
   final Duration              timeout;
-  final bool                  printHeaders;
-  final bool                  printRequest;
+  final bool                  silent;
+  final bool                  printUrl;
+  final bool                  printHeader;
+  final bool                  printBody;
   final bool                  printResponse;
+  final Function()?           onStart;
+  final Function()?           onFinish;
 
   NRequest(this.url, {
     this.body          = const <String, dynamic>{},
     this.files         = const [],
     this.timeout       = const Duration(minutes: 5),
-    this.printHeaders  = false,
-    this.printRequest  = false,
+    this.silent        = false,
+    this.printUrl      = false,
+    this.printHeader   = false,
+    this.printBody     = false,
     this.printResponse = false,
     this.headers,
     this.token,
+    this.onStart,
+    this.onFinish,
   });
 
   Future<ResponseData> _request(RequestType type) async => await NCustomRequest.make(
@@ -42,17 +49,19 @@ class NRequest{
     files         : files,
     timeout       : timeout,
     token         : token,
-    printHeaders  : printHeaders,
-    printRequest  : printRequest,
+    printUrl      : printUrl,
+    silent        : silent,
+    printHeader   : printHeader,
+    printBody     : printBody,
     printResponse : printResponse,
   ).then((value) => value);
 
-  Future<R> Download  <R>(Function(Uint8List? data) onValue) async => await NCustomRequest.download(
+  Future<R> download  <R>(Function(Uint8List? data) onValue) async => await NCustomRequest.download(
     url           : url,
     headers       : headers,
     timeout       : timeout,
     token         : token,
-    printHeaders  : printHeaders,
+    printHeader   : printHeader,
   ).then((value) => onValue.call(value));
 
   Future<R> get       <R>(Function(ResponseData response) onValue) async => await _request(RequestType.get   ).then((response) async => await onValue.call(response) );