+++ test/n_request_test.dart
@@ -1,5 +1,15 @@
-import 'package:flutter_test/flutter_test.dart';
+import 'package:n_request/n_request.dart';
 
-void main() {
-  test('adds one to input values', () {});
+void main() async {
+  await NRequest(
+    "https://pokeapi.co/api/v2/ability/?limit=20&offset=20",
+    headers       : {"testKey": "testValue"},
+    body          : {"testBody": "testBody"},
+    onStart       : ()=> print("start"),
+    onFinish      : ()=> print("finish"),
+    printUrl      : true,
+    printHeader   : true,
+    printBody     : true,
+    printResponse : true
+  ).get((response) {});
 }